# api/swagger.yaml
openapi: 3.0.1
info:
  title: Traffic Jam API
  description: API to manage traffic jams
  version: 1.0.0
servers:
  - url: https://localhost:44377/API/V1
    description: localhost (what is this for?)
  - url: https://virtserver.swaggerhub.com/peterjochum/traffic-jam_api/1.0.0
    description: SwaggerHub API Auto Mocking
paths:
  /trafficjam:
    get:
      summary: Get all traffic jams
      operationId: getAllTrafficJams
      responses:
        "200":
          description: OK
    put:
      summary: Update traffic jam report
      operationId: putTrafficJam
      requestBody:
        $ref: "#/components/requestBodies/TrafficJam"
      responses:
        "404":
          description: Traffic jam was not found
    post:
      summary: Create a new traffic jam report
      operationId: addTrafficJam
      requestBody:
        $ref: "#/components/requestBodies/TrafficJam"
      responses:
        "400":
          description: Malformed request - included all required fields?
  /trafficjam/{id}:
    get:
      summary: Return traffic jam by {id}
      operationId: getTrafficJam
      parameters:
        - name: id
          in: path
          description: ID of traffic jam to return
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "404":
          description: Traffic jam was not found
    delete:
      summary: delete traffic jam with {id}
      operationId: deleteTrafficJam
      parameters:
        - name: id
          in: path
          description: ID of the traffic jam to delete
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "404":
          description: Traffic jam was not found
components:
  schemas:
    TrafficJam:
      required:
        - durationInSeconds
        - id
        - latitude
        - longitude
      type: object
      properties:
        id:
          type: integer
          format: int64
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
        durationInSeconds:
          type: integer
          format: int32
  responses:
    TrafficJamNotFound:
      description: Traffic jam was not found
    TrafficJamBadRequest:
      description: Malformed request - included all required fields?
  requestBodies:
    TrafficJam:
      description: Traffic jam that needs to be added to the store
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TrafficJam"
        application/xml:
          schema:
            $ref: "#/components/schemas/TrafficJam"
      required: true